requirements-multimodal.txt
openai==1.51.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy>=1.22.0,<1.27.0
pyaudio==0.2.14
speechrecognition==3.10.4
pydub==0.25.1
TTS==0.21.0
torch>=2.0.0
torchaudio>=2.0.0
requests==2.31.0
python-dotenv==1.0.1
librosa>=0.10.0
scipy>=1.10.0